task 2 -fix

+++ 0x00-python_variable_annotations/2-floor.py
@@ -8,4 +8,4 @@ def floor(n: float) -> int:
     n: float
     return: int
     """
-    return floor(n)
+    return int(floor(n))
